+++ Assets/Packages/com.studio23.ss2.interactionsystem/Runtime/Core/InspectionManager.cs
@@ -34,6 +34,7 @@ namespace Studio23.SS2.InteractionSystem.Core
         public bool ExaminationYAxisInverted = false;
 
         [SerializeField] Camera _mainCamera;
+        [SerializeField] private bool _setMainCameraAsInspectCamera;
         [SerializeField] Camera  _inspectionCamera;
         [SerializeField] private Canvas _inspectionBackgroundCanvas;
 
@@ -53,7 +54,11 @@ namespace Studio23.SS2.InteractionSystem.Core
             InteractionManager.Instance.InputHandler.InspectionDragButton.AddCancelledCallback(gameObject, HandleInspectDragCancelled);
             InteractionManager.Instance.InputHandler.InteractCancelButton.AddPerformedCallback(gameObject, HandleInspectionCancelled);
             InteractionManager.Instance.InputHandler.InspectResetButton.AddPerformedCallback(gameObject, HandleInspectResetPerformed);
+
+            
         }
+
+
         
         public void HandleInspectResetPerformed()
         {
@@ -180,8 +185,8 @@ namespace Studio23.SS2.InteractionSystem.Core
             _ogOrientation = _examinationObject.localRotation;
 
             _examinationObject.gameObject.SetActive(true);
-            _examinationObject.parent = InspectionObjectParent;
-            _examinationObject.localPosition = inspectable.InspectionPosOffset;
+            //_examinationObject.parent = InspectionObjectParent;
+            _examinationObject.localPosition = Vector3.zero;
             _examinationObject.localRotation = Quaternion.Euler(inspectable.InspectionRotOffset);
             _examinationObject.localScale = inspectable.InspectionScale;
 
@@ -200,7 +205,7 @@ namespace Studio23.SS2.InteractionSystem.Core
             _inspectionBackgroundCanvas.worldCamera = _mainCamera;
             _inspectionBackgroundCanvas.gameObject.SetActive(true);
             //urp camera stacking
-            ActivateInspectionCamStacking();
+            //ActivateInspectionCamStacking();
 
             _wantsToCancel = false;
             _isDragging = false;
@@ -224,6 +229,14 @@ namespace Studio23.SS2.InteractionSystem.Core
             {
                 _mainCamera = Camera.main;    
             }
+
+            if(_setMainCameraAsInspectCamera)
+                _inspectionCamera = _mainCamera;
+        }
+
+        public void SetInspectionCamera(Camera inspectionCamera)
+        {
+            _inspectionCamera = inspectionCamera;
         }
 
         public void HandleInspectablePaused(InspectableBase inspectable)
@@ -240,7 +253,7 @@ namespace Studio23.SS2.InteractionSystem.Core
             UnMoveInspectableForInspection(inspectable);
             
             _inspectionBackgroundCanvas.gameObject.SetActive(false);
-            DeactivateInspectionCamStacking();
+            //DeactivateInspectionCamStacking();
 
             _curInspectable = null;
         }